feat(react-spreadsheet): VirtualSpreadsheet no longer needs a type parameter
Renamed the existing VirtualSpreadsheet and VirtualSpreadsheetProps to VirtualSpreadsheetGeneric and VirtualSpreadsheetGenericProps
Added new implementations of VirtualSpreadsheet and VirtualSpreadsheetProps that are instantiations of the generic versions with unknown.
Configured ESLint to allow empty interfaces that extend another interface.

BREAKING CHANGE: Code that used VirtualSpreadsheet or VirtualSpreadsheetProps with a type parameter needs to either remove the parameter or switch to the generic version.